fix(backend): 修复mongo增加分片提单报float不能interpreted成int的问题 #13420

+++ dbm-ui/backend/ticket/builders/mongodb/mongo_add_shard.py
@@ -60,16 +60,16 @@ class MongoDBAddShardResourceParamBuilder(BaseMongoDBOperateResourceParamBuilder
         # 资源申请的一些参数补充
         self.patch_info_common_affinity(role="mongodb", remain_machine_type=MachineType.MONGODB, tolerance=0.5)
         for info in self.ticket_data["infos"]:
-            add_group = info["add_shards_num"] / info["node_replicaset_count"]  # 新增组数
+            add_group = int(info["add_shards_num"] / info["node_replicaset_count"])  # 新增组数
             origin_group_info = info["resource_spec"].pop("mongodb")
-            origin_group_info["count"] = origin_group_info["count"] / add_group
+            origin_group_info["count"] = int(origin_group_info["count"] / add_group)
             for i in range(add_group):
                 info["resource_spec"][f"mongodb_{i}"] = origin_group_info
 
     def post_callback(self):
         with self.next_flow_manager() as next_flow:
             for info in next_flow.details["ticket_data"]["infos"]:
-                add_group = info["add_shards_num"] / info["node_replicaset_count"]  # 新增组数
+                add_group = int(info["add_shards_num"] / info["node_replicaset_count"])  # 新增组数
                 info["mongo_add_shards"] = []
                 for i in range(add_group):
                     info["mongo_add_shards"] = info["mongo_add_shards"].append(info.pop(f"mongodb_{i}"))
